Use recipeList rather than option list of named predicates
