+++ gradle/licenseHeader.txt
@@ -1,4 +1,4 @@
-Copyright 2021 the original author or authors.
+Copyright 2022 the original author or authors.
 <p>
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
