--- src/main/java/org/openrewrite/java/security/search/secret/ArtifactorySecretPredicateGroup.java
@@ -1,35 +1,0 @@
-/*
- * Copyright 2021 the original author or authors.
- * <p>
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * <p>
- * https://www.apache.org/licenses/LICENSE-2.0
- * <p>
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.openrewrite.java.security.search.secret;
-
-import org.openrewrite.ExecutionContext;
-
-import java.util.Arrays;
-import java.util.List;
-
-class ArtifactorySecretPredicateGroup implements SecretPredicateGroup {
-    @Override
-    public String getName() {
-        return "Artifactory";
-    }
-
-    @Override
-    public List<SecretPredicate<String, String, ExecutionContext>> secretPredicates() {
-        return Arrays.asList(new SecretRegexPredicate(null, "(?:\\s|=|:|\"|^)AP[\\dABCDEF][a-zA-Z0-9]{8,}(?:\\s|\"|$)"),
-                new SecretRegexPredicate(null, "(?:\\s|=|:|\"|^)AKC[a-zA-Z0-9]{10,}(?:\\s|\"|$)")
-        );
-    }
-}
