+++ src/main/java/org/openrewrite/java/security/secrets/package-info.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 the original author or authors.
+ * Copyright 2022 the original author or authors.
  * <p>
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,6 +14,6 @@
  * limitations under the License.
  */
 @NonNullApi
-package org.openrewrite.java.security.search.secret;
+package org.openrewrite.java.security.secrets;
 
 import org.openrewrite.internal.lang.NonNullApi;
